requirements.txt
PyQt6>=6.5
PyMuPDF>=1.23
python-docx>=1.1
paddleocr>=2.7,<3.0
paddlepaddle>=2.5,<3.0
Pillow>=10.0
numpy>=1.24
